move common code to directory to add to path
