+++ examples/advanced/experiment-tracking/mlflow/jobs/hello-pt-tb-mlflow/app/config/config_fed_server.json
@@ -12,7 +12,7 @@
       "path": "nvflare.app_common.pt.pt_file_model_persistor.PTFileModelPersistor",
       "args": {
         "model": {
-          "path": "simple_network.SimpleNetwork"
+          "path": "pt.simple_network.SimpleNetwork"
         }
       }
     },
@@ -49,16 +49,16 @@
       "path": "nvflare.app_opt.tracking.mlflow.mlflow_receiver.MLflowReceiver",
       "args": {
         "kwargs": {
-          "experiment_name": "hello-pt-experiment3",
-          "run_name": "hello-pt-with-mlflow1",
+          "experiment_name": "hello-pt-experiment",
+          "run_name": "hello-pt-with-mlflow",
           "experiment_tags": {
             "mlflow.note.content": "## **Hello PyTorch experiment with MLflow**"
           },
           "run_tags": {
-            "mlflow.note.content": "## Federated Experiment tracking with MLfLow \n### Example of using **[NVIDIA FLARE](https://nvflare.readthedocs.io/en/main/index.html)** to train an image classifier using federated averaging ([FedAvg]([FedAvg](https://arxiv.org/abs/1602.05629))) and [PyTorch](https://pytorch.org/) as the deep learning training framework. This example also highlights the Flare streaming capability from the clients to the server for server delivery to MLflow.\n\n> **_NOTE:_** \n This example uses the *[CIFAR-10](https://www.cs.toronto.edu/~kriz/cifar.html)* dataset and will load its data within the trainer code.\n"
+            "mlflow.note.content": "## Federated Experiment tracking with MLflow \n### Example of using **[NVIDIA FLARE](https://nvflare.readthedocs.io/en/main/index.html)** to train an image classifier using federated averaging ([FedAvg]([FedAvg](https://arxiv.org/abs/1602.05629))) and [PyTorch](https://pytorch.org/) as the deep learning training framework. This example also highlights the Flare streaming capability from the clients to the server for server delivery to MLflow.\n\n> **_NOTE:_** \n This example uses the *[CIFAR-10](https://www.cs.toronto.edu/~kriz/cifar.html)* dataset and will load its data within the trainer code.\n"
           }
         },
-        "artifact_location": "artifacts2"
+        "artifact_location": "artifacts"
       }
     }
   ],
