+++ examples/advanced/experiment-tracking/mlflow/jobs/hello-pt-tb-mlflow/app/custom/pt_learner.py
@@ -15,8 +15,8 @@
 import os.path
 
 import torch
-from pt_constants import PTConstants
-from simple_network import SimpleNetwork
+from pt.pt_constants import PTConstants
+from pt.simple_network import SimpleNetwork
 from torch import nn
 from torch.optim import SGD
 from torch.utils.data.dataloader import DataLoader
